fix ClangCL build [-Werror,-Wdouble-promotion]
D:\a\sentry-native\sentry-native\src\sentry_envelope.c(395,26): error : implicit conversion increases floating-point precision: 'float' to 'double' [-Werror,-Wdouble-promotion] [C:\Users\runneradmin\AppData\Local\Temp\pytest-of-runneradmin\pytest-0\cmake190\sentry.vcxproj]
  C:\Program Files (x86)\Windows Kits\10\Include\10.0.26100.0\ucrt\corecrt_math.h(111,13): note: expanded from macro 'NAN'
  C:\Program Files (x86)\Windows Kits\10\Include\10.0.26100.0\ucrt\corecrt_math.h(105,20): note: expanded from macro '_UCRT_NAN'

+++ src/sentry_envelope.c
@@ -392,7 +392,7 @@ sentry__envelope_add_event(sentry_envelope_t *envelope, sentry_value_t event)
         traces_sample_rate = options->traces_sample_rate;
     }
     sentry_value_t dsc = sentry_value_new_null();
-    double sample_rand = NAN;
+    double sample_rand = (double)NAN;
     SENTRY_WITH_SCOPE (scope) {
         dsc = sentry__value_clone(scope->dynamic_sampling_context);
         sample_rand = sentry_value_as_double(sentry_value_get_by_key(
